configs: New config in init for v5.4-rc1
Message-id: <20191014162300.16308-19-jcline@redhat.com>
Patchwork-id: 278894
O-Subject: [ARK INTERNAL PATCH] configs: New config in init for v5.4-rc1
Bugzilla:
RH-Acked-by: Laura Abbott <labbott@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* init/Kconfig (help): MODULE_ALLOW_MISSING_NAMESPACE_IMPORTS (bool)
    Symbols exported with EXPORT_SYMBOL_NS*() are considered exported in
    a namespace. A module that makes use of a symbol exported with such a
    namespace is required to import the namespace via MODULE_IMPORT_NS().
    There is no technical reason to enforce correct namespace imports,
    but it creates consistency between symbols defining namespaces and
    users importing namespaces they make use of. This option relaxes this
    requirement and lifts the enforcement when loading a module.

    If unsure, say N.

Signed-off-by: Jeremy Cline <jcline@redhat.com>