[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BACKLIGHT_QCOM_WLED:

 If you have the Qualcomm PMIC, say Y to enable a driver for the
 WLED block. Currently it supports PM8941 and PMI8998.

 Symbol: BACKLIGHT_QCOM_WLED [=n]
 Type  : tristate
 Prompt: Qualcomm PMIC WLED Driver
   Location:
     -> Device Drivers
       -> Graphics support
         -> Backlight & LCD device support
           -> Lowlevel Backlight controls (BACKLIGHT_CLASS_DEVICE [=y])
   Defined at drivers/video/backlight/Kconfig:285
   Depends on: HAS_IOMEM [=y] && BACKLIGHT_CLASS_DEVICE [=y]
   Selects: REGMAP [=y]

---