[redhat] New configs in drivers/ptp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PTP_1588_CLOCK_INES:

 This driver adds support for using the ZHAW InES 1588 IP
 core.  This clock is only useful if the MII bus of your MAC
 is wired up to the core.

 Symbol: PTP_1588_CLOCK_INES [=n]
 Type  : tristate
 Prompt: ZHAW InES PTP time stamping IP core
   Location:
     -> Device Drivers
       -> PTP clock support
   Defined at drivers/ptp/Kconfig:78
   Depends on: NETWORK_PHY_TIMESTAMPING [=y] && PHYLIB [=y] && PTP_1588_CLOCK [=y]

---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>