Fix some awkward edge cases in gen_config_patches.sh
This refactors the awk script that sorts new configuration symbols by
subsystem. There were two awkward edge cases.

The first was in the handling of the drivers/{net,net/ethernet} that
failed to strip the trailing ":<linenumber>" from new configuration
symbols in drivers/net/{Kconfig,ethernet/Kconfig}. This led to branch
names that included the ":<linenumber>" which aren't valid branch names
and caused the script to create commits in incorrect branches.

The second edge case was that it replaced "/" with "_" in the filename
it created. The problem with this is that some subsystem directories
have "_" in the name (e.g. net/vmw_vsock) When it turned the filename
back into a kernel tree path it would mangle the name and fail to find
any maintainers for it. The new placeholder character is ":" which is
not currently in any kernel directory name and seems unlikely to appear
in one.

Additionally, there are some changes for readability.

Signed-off-by: Jeremy Cline <jcline@redhat.com>