New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CHELSIO_TLS_DEVICE:

 This flag enables support for kernel tls offload over Chelsio T6
 crypto accelerator. CONFIG_CHELSIO_TLS_DEVICE flag can be enabled
 only if CONFIG_TLS and CONFIG_TLS_DEVICE flags are enabled.

 Symbol: CHELSIO_TLS_DEVICE [=y]
 Type  : bool
 Defined at drivers/crypto/chelsio/Kconfig:46
   Prompt: Chelsio Inline KTLS Offload
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && CHELSIO_T4 [=m] && TLS_DEVICE [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
 Selects: CRYPTO_DEV_CHELSIO [=m]

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>

V2: Update CONFIG_CHELSIO_TLS_DEVICE to =m because inline is not allowed.