[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HID_GLORIOUS:

 Support for Glorious PC Gaming Race mice such as
 the Glorious Model O, O- and D.

 Symbol: HID_GLORIOUS [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:365
   Prompt: Glorious PC Gaming Race mice
   Depends on: INPUT [=y] && HID [=y]
   Location:
     -> Device Drivers
       -> HID support
         -> HID bus support (HID [=y])
           -> Special HID drivers

---

 CONFIG_HID_MCP2221:

 Provides I2C and SMBUS host adapter functionality over USB-HID
 through MCP2221 device.

 To compile this driver as a module, choose M here: the module
 will be called hid-mcp2221.ko.

 Symbol: HID_MCP2221 [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:1155
   Prompt: Microchip MCP2221 HID USB-to-I2C/SMbus host support
   Depends on: INPUT [=y] && HID [=y] && USB_HID [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> HID support
         -> HID bus support (HID [=y])
           -> Special HID drivers

---

Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>