[redhat] New configs in drivers/mmc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MMC_HSQ:

 This selects the MMC Host Software Queue support. This may increase
 performance, if the host controller and its driver supports it.

 If you have a controller/driver supporting this interface, say Y or M
 here.

 If unsure, say N.

 Symbol: MMC_HSQ [=n]
 Type  : tristate
 Defined at drivers/mmc/host/Kconfig:953
   Prompt: MMC Host Software Queue support
   Depends on: MMC [=m]
   Location:
     -> Device Drivers
       -> MMC/SD/SDIO card support (MMC [=m])
 Selected by [n]:
   - MMC_SDHCI_SPRD [=n] && MMC [=m] && ARCH_SPRD && MMC_SDHCI_PLTFM [=m]

---

Cc: Gopal Tiwari <gtiwari@redhat.com>