New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_LEDS_AW2013:

 This option enables support for the AW2013 3-channel
 LED driver.

 To compile this driver as a module, choose M here: the module
 will be called leds-aw2013.

 Symbol: LEDS_AW2013 [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:106
   Prompt: LED support for Awinic AW2013
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])

Cc: Tony Camuso <tcamuso@redhat.com>