New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MFD_MP2629:

 Select this option to enable support for Monolithic Power Systems
 battery charger. This provides ADC, thermal and battery charger power
 management functions.

 Symbol: MFD_MP2629 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:452
   Prompt: Monolithic Power Systems MP2629 ADC and Battery charger
   Depends on: HAS_IOMEM [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
 Selects: REGMAP_I2C [=m]

Cc: Torez Smith <torez@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>