New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_EFI_CUSTOM_SSDT_OVERLAYS:

 Allow loading of an ACPI SSDT overlay from an EFI variable specified
 by a kernel command line option.

 See Documentation/admin-guide/acpi/ssdt-overlays.rst for more
 information.

 Symbol: EFI_CUSTOM_SSDT_OVERLAYS [=y]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:282
   Prompt: Load custom ACPI SSDT overlay from an EFI variable
   Depends on: EFI_VARS [=y] && ACPI [=y]
   Location:
     -> Firmware Drivers

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>