[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_SMCCC_SOC_ID:

 Include support for the SoC bus on the ARM SMCCC firmware based
 platforms providing some sysfs information about the SoC variant.

 Symbol: ARM_SMCCC_SOC_ID [=y]
 Type  : bool
 Defined at drivers/firmware/smccc/Kconfig:18
   Prompt: SoC bus device for the ARM SMCCC SOC_ID
   Depends on: HAVE_ARM_SMCCC_DISCOVERY [=y]
   Location:
     -> Firmware Drivers
 Selects: SOC_BUS [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>