[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_RCU_REF_SCALE_TEST:

 This option provides a kernel module that runs performance tests
 useful comparing RCU with various read-side synchronization mechanisms.
 The kernel module may be built after the fact on the running kernel to be
 tested, if desired.

 Say Y here if you want these performance tests built into the kernel.
 Say M if you want to build it as a module instead.
 Say N if you are unsure.

 Symbol: RCU_REF_SCALE_TEST [=n]
 Type  : tristate
 Defined at kernel/rcu/Kconfig.debug:64
   Prompt: Scalability tests for read-side synchronization (RCU and others)
   Depends on: DEBUG_KERNEL [=y]
   Location:
     -> Kernel hacking
       -> RCU Debugging
 Selects: TORTURE_TEST [=n] && SRCU [=y] && TASKS_RCU [=n] && TASKS_RUDE_RCU [=y] && TASKS_TRACE_RCU [=n]

---

Cc: Waiman Long <longman@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>