[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SENSORS_CORSAIR_CPRO:

 If you say yes here you get support for the Corsair Commander Pro
 controller.

 This driver can also be built as a module. If so, the module
 will be called corsair-cpro.

 Symbol: SENSORS_CORSAIR_CPRO [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:442
   Prompt: Corsair Commander Pro controller
   Depends on: HWMON [=y] && HID [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])

---

Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>