[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_THERMAL_NETLINK:

 The thermal framework has a netlink interface to do thermal
 zones discovery, temperature readings and events such as
 trip point crossed, cooling device update or governor
 change. It is recommended to enable the feature.

 Symbol: THERMAL_NETLINK [=n]
 Type  : bool
 Defined at drivers/thermal/Kconfig:20
   Prompt: Thermal netlink management
   Depends on: THERMAL [=y] && NET [=y]
   Location:
     -> Device Drivers
       -> Thermal drivers (THERMAL [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>