Filter out upstream commits from changelog
I updated the release code to include more changelog info
in the os-build branch.  This resulted in adding upstream
merge commits.  This is not interesting for a Fedora tree
as it is covered by a generic 'merge' entry.

Use a simple git trick to filter those upstream commits out, "^master".
See 'man gitrevisions' for more tricks.

Signed-off-by: Don Zickus <dzickus@redhat.com>