[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_FB_ARMCLCD:

 This framebuffer device driver is for the ARM PrimeCell PL110
 Colour LCD controller.  ARM PrimeCells provide the building
 blocks for System on a Chip devices.

 If you want to compile this as a module (=code which can be
 inserted into and removed from the running kernel), say M
 here and read <file:Documentation/kbuild/modules.rst>.  The module
 will be called amba-clcd.

 Symbol: FB_ARMCLCD [=n]
 Type  : tristate
 Defined at drivers/video/fbdev/Kconfig:275
   Prompt: ARM PrimeCell PL110 support
   Depends on: (ARM || ARM64 [=y] || COMPILE_TEST [=n]) && FB [=y] && ARM_AMBA [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Frame buffer Devices
           -> Support for frame buffer devices (FB [=y])
 Selects: FB_CFB_FILLRECT [=y] && FB_CFB_COPYAREA [=y] && FB_CFB_IMAGEBLIT [=y] && FB_MODE_HELPERS [=n] && VIDEOMODE_HELPERS [=n] && BACKLIGHT_CLASS_DEVICE [=y]

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>