[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PPC_RTAS_FILTER:

 The RTAS syscall API has security issues that could be used to
 compromise system integrity. This option enforces restrictions on the
 RTAS calls and arguments passed by userspace programs to mitigate
 these issues.

 Say Y unless you know what you are doing and the filter is causing
 problems for you.

 Symbol: PPC_RTAS_FILTER [=y]
 Type  : bool
 Defined at arch/powerpc/Kconfig:991
   Prompt: Enable filtering of RTAS syscalls
   Depends on: PPC_RTAS [=y]
   Location:
     -> Kernel options

---

Cc: kernel-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>