[redhat] New configs in fs/fuse
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_FUSE_DAX:

 This allows bypassing guest page cache and allows mapping host page
 cache directly in guest address space.

 If you want to allow mounting a Virtio Filesystem with the "dax"
 option, answer Y.

 Symbol: FUSE_DAX [=y]
 Type  : bool
 Defined at fs/fuse/Kconfig:42
   Prompt: Virtio Filesystem Direct Host Memory Access support
   Depends on: VIRTIO_FS [=m] && FS_DAX [=y] && DAX_DRIVER [=y]
   Location:
     -> File systems
       -> FUSE (Filesystem in Userspace) support (FUSE_FS [=m])
         -> Virtio Filesystem (VIRTIO_FS [=m])
 Selects: INTERVAL_TREE [=y]

---

Cc: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>