[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SERIAL_BCM63XX:

 This enables the driver for the onchip UART core found on
 the following chipsets:

   BCM33xx (cable modem)
   BCM63xx/BCM63xxx (DSL)
   BCM68xx (PON)
   BCM7xxx (STB) - DOCSIS console

 Symbol: SERIAL_BCM63XX [=n]
 Type  : tristate
 Defined at drivers/tty/serial/Kconfig:1117
   Prompt: Broadcom BCM63xx/BCM33xx UART support
   Depends on: TTY [=y] && HAS_IOMEM [=y] && COMMON_CLK [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
 Selects: SERIAL_CORE [=y]

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>