[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARCH_BCM4908:

 This enables support for the Broadcom BCM4906, BCM4908 and
 BCM49408 SoCs. These SoCs use Brahma-B53 cores and can be
 found in home routers.

 Symbol: ARCH_BCM4908 [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:46
   Prompt: Broadcom BCM4908 family
   Location:
     -> Platform selection
 Selects: GPIOLIB [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>