[redhat] New configs in drivers/md
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DM_MULTIPATH_IOA:

 This path selector selects the path based on the CPU the IO is
 executed on and the CPU to path mapping setup at path addition time.

 If unsure, say N.

 Symbol: DM_MULTIPATH_IOA [=n]
 Type  : tristate
 Defined at drivers/md/Kconfig:466
   Prompt: I/O Path Selector based on CPU submission
   Depends on: MD [=y] && DM_MULTIPATH [=m]
   Location:
     -> Device Drivers
       -> Multiple devices driver support (RAID and LVM) (MD [=y])
         -> Device mapper support (BLK_DEV_DM [=m])
           -> Multipath target (DM_MULTIPATH [=m])

---

Cc: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>