rpmspec: ship gpio-watch.debug in the proper debuginfo package
The gpio-watch tool is shipped within the kernel-tools package. However,
when it was added some time ago, the debuginfo pattern inside the spec
file was not updated, so the gpio-watch.debug file ends up in the
kernel-debuginfo-common package instead of the kernel-tools-debuginfo.

This change fixes the problem, and also it's likely will fix an
update/downgrade test done on CI for RHEL that reports:

Error: Transaction test error:
  file /usr/lib/debug/usr/bin/gpio-watch.debug from install of
kernel-debuginfo-common-x86_64-5.11.0-0.rc7.151.el9.x86_64 conflicts
with file from package kernel-debuginfo-common-x86_64-5.11.0-1.el9.x86_64

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>