[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_SMMU_V3_SVA:

 Support for sharing process address spaces with devices using the
 SMMUv3.

 Say Y here if your system supports SVA extensions such as PCIe PASID
 and PRI.

 Symbol: ARM_SMMU_V3_SVA [=n]
 Type  : bool
 Defined at drivers/iommu/Kconfig:311
   Prompt: Shared Virtual Addressing support for the ARM SMMUv3
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU_V3 [=y]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> ARM Ltd. System MMU Version 3 (SMMUv3) Support (ARM_SMMU_V3 [=y])

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>