[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DEBUG_ENTRY:

 This option enables sanity checks in s390 low-level entry code.
 Some of these sanity checks may slow down kernel entries and
 exits or otherwise impact performance.

 If unsure, say N.

 Symbol: DEBUG_ENTRY [=n]
 Type  : bool
 Defined at arch/s390/Kconfig.debug:9
   Prompt: Debug low-level entry code
   Depends on: DEBUG_KERNEL [=y]
   Location:
     -> Kernel hacking
       -> s390 Debugging

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- Set CONFIG_DEBUG_ENTRY for s390x debug variant.
- Removed redundant ark/generic CONFIG_DEBUG_ENTRY file.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>