[redhat] New configs in drivers/net/ethernet/mellanox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MLX5_SF:

 Build support for subfuction device in the NIC. A Mellanox subfunction
 device can support RDMA, netdevice and vdpa device.
 It is similar to a SRIOV VF but it doesn't require SRIOV support.

 Symbol: MLX5_SF [=n]
 Type  : bool
 Defined at drivers/net/ethernet/mellanox/mlx5/core/Kconfig:205
   Prompt: Mellanox Technologies subfunction device support using auxiliary device
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MELLANOX [=y] && MLX5_CORE [=m] && MLX5_CORE_EN [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Mellanox devices (NET_VENDOR_MELLANOX [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>