[redhat] New configs in drivers/scsi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SCSI_HISI_SAS_DEBUGFS_DEFAULT_ENABLE:

 Set Y to default enable DEBUGFS for SCSI_HISI_SAS

 Symbol: SCSI_HISI_SAS_DEBUGFS_DEFAULT_ENABLE [=n]
 Type  : bool
 Defined at drivers/scsi/hisi_sas/Kconfig:22
   Prompt: HiSilicon SAS debugging default enable
   Depends on: SCSI_LOWLEVEL [=y] && SCSI [=y] && SCSI_HISI_SAS [=m]
   Location:
     -> Device Drivers
       -> SCSI device support
         -> SCSI low-level drivers (SCSI_LOWLEVEL [=y])
           -> HiSilicon SAS (SCSI_HISI_SAS [=m])

---

Cc: "Ewan D. Milne" <emilne@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>