[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HARDLOCKUP_DETECTOR:

 Say Y here to enable the kernel to act as a watchdog to detect
 hard lockups.

 Hardlockups are bugs that cause the CPU to loop in kernel mode
 for more than 10 seconds, without letting other interrupts have a
 chance to run.  The current stack trace is displayed upon detection
 and the system will stay locked up.

 Symbol: HARDLOCKUP_DETECTOR [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1000
   Prompt: Detect Hard Lockups
   Depends on: DEBUG_KERNEL [=y] && !S390 && (HAVE_HARDLOCKUP_DETECTOR_PERF [=y] || HAVE_HARDLOCKUP_DETECTOR_ARCH [=n])
   Location:
     -> Kernel hacking
       -> Debug Oops, Lockups and Hangs
 Selects: LOCKUP_DETECTOR [=y] && HARDLOCKUP_DETECTOR_PERF [=n] && HARDLOCKUP_DETECTOR_ARCH

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>