[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_NVIDIA_CARMEL_CNP_ERRATUM:

 If CNP is enabled on Carmel cores, non-sharable TLBIs on a core will not
 invalidate shared TLB entries installed by a different core, as it would
 on standard ARM cores.

 If unsure, say Y.

 Symbol: NVIDIA_CARMEL_CNP_ERRATUM [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:813
   Prompt: NVIDIA Carmel CNP: CNP on Carmel semantically different than ARM cores
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>