Use upstream/master for merge-base with fallback to master
In Makefile.common there is logic that determines what the
merge-base is and is used to generate changelogs correctly.

The logic assumes a 'master' branch exists and is up to date.
However, the documentation recommends using 'upstream/master' to
keep everything up to date.

Let's sync the documentation and code and use upstream/master if
it exists otherwise, use 'master' as the fallback.  If neither
exists error out.

Update redhat/genspec.sh to handle a different master too.

The 'error' out solves another problem when a master branch is not
defined not is 'upstream' used as recommended.

V2: change 'upstream' to 'origin' (Justin F.)

Signed-off-by: Don Zickus <dzickus@redhat.com>