[redhat] New configs in drivers/net/can
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CAN_ETAS_ES58X:

 This driver supports the ES581.4, ES582.1 and ES584.1 interfaces
 from ETAS GmbH (https://www.etas.com/en/products/es58x.php).

 To compile this driver as a module, choose M here: the module
 will be called etas_es58x.

 Symbol: CAN_ETAS_ES58X [=n]
 Type  : tristate
 Defined at drivers/net/can/usb/Kconfig:23
   Prompt: ETAS ES58X CAN/USB interfaces
   Depends on: NET [=y] && CAN [=m] && CAN_DEV [=m] && USB [=y]
   Location:
     -> Networking support (NET [=y])
       -> CAN bus subsystem support (CAN [=m])
         -> CAN Device Drivers
           -> Platform CAN drivers with Netlink support (CAN_DEV [=m])
             -> CAN USB interfaces
 Selects: CRC16 [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>