[redhat] New configs in fs/zonefs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ZONEFS_FS:

 zonefs is a simple file system which exposes zones of a zoned block
 device (e.g. host-managed or host-aware SMR disk drives) as files.

 If unsure, say N.

 Symbol: ZONEFS_FS [=n]
 Type  : tristate
 Defined at fs/zonefs/Kconfig:1
   Prompt: zonefs filesystem support
   Depends on: BLOCK [=y] && BLK_DEV_ZONED [=y]
   Location:
     -> File systems
 Selects: FS_IOMAP [=y] && CRC32 [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>