CI: Finish up c9s config
The configuration for c9s was, up till now, a commented out placeholder.
The plans are finished now so we can also update the config accordingly.

We want two pipelines:

- A full c9s pipeline (build and test), publicly available
- A build for RHEL9 (RH internal only), to ensure nothing breaks (this
  should not happen, but better be safe)

Add a new anchor for RHEL build-only and the extra pipelines. As we now
have a lot of pipelines defined in this file, add comments to mark where
the actual definitions start for easier orientation.

Bugzilla: INTERNAL
Upstream Status: RHEL-only

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>