Build CRYPTO_SHA3_*_S390 inline for s390 zfcpdump
While the configs for CRYPTO_SHA3_512_S390 and CRYPTO_SHA3_256_S390 were
reviewed and set to a module, the zfcpdump kernel does not support
modules. We have to either build them inline (as we have been for a
while) or we need to turn them off all together for the zfcpdump kernel.
This MR exists because it was created in pending and got dropped without
a proper MR to review and decide on the proper solution. I have set them
to build inline, but am happy to change them based on review.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>