Change s390x CONFIG_NODES_SHIFT from 4 to 1
Upstream change 701dc81e7412d removed the fake numa support on s390.  As
a result, though some generic structure was left in place for
applications that expect NUMA interfaces to exist. The current reviewed
and approved setting for NODES_SHIFT on s390 is 4, but that is no longer
valid.  The only valid options are 1 or turning off numa completely
which can be problematic for some applications.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>