[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KEYBOARD_CYPRESS_SF [=n]
 Type  : tristate
 Defined at drivers/input/keyboard/Kconfig:791
   Prompt: Cypress StreetFighter touchkey support
   Depends on: INPUT [=y] && INPUT_KEYBOARD [=y] && I2C [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Input device support
           -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
             -> Keyboards (INPUT_KEYBOARD [=y])

---

Cc: Tony Camuso <tcamuso@redhat.com>
Cc: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>