New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_CHACHA_S390 [=n]
 Type  : tristate
 Defined at drivers/crypto/Kconfig:216
   Prompt: ChaCha20 stream cipher
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && S390 [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
         -> Hardware crypto devices (CRYPTO_HW [=y])
 Selects: CRYPTO_ALGAPI [=y] && CRYPTO_SKCIPHER [=y] && CRYPTO_CHACHA20 [=m]

Signed-off-by: Vladis Dronov <vdronov@redhat.com>