[redhat] New configs in drivers/net/ethernet/vertexcom
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NET_VENDOR_VERTEXCOM [=n]
 Type  : bool
 Defined at drivers/net/ethernet/vertexcom/Kconfig:6
   Prompt: Vertexcom devices
   Depends on: NETDEVICES [=y] && ETHERNET [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Ethernet driver support (ETHERNET [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>