[redhat] New configs in drivers/char
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HW_RANDOM_CN10K [=y]
 Type  : tristate
 Defined at drivers/char/hw_random/Kconfig:541
   Prompt: Marvell CN10K Random Number Generator support
   Depends on: HW_RANDOM [=y] && PCI [=y] && ARM64 [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Character devices
           -> Hardware Random Number Generator Core support (HW_RANDOM [=y])

---

Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>