configs/fedora: Enable CONFIG_NFC_PN532_UART for use PN532 NFC module
PN532 is a cost effective NFC module. In order to use PN532 with pcsc
etc in a Linux-based PC with UART, pn532_uart module is required. This
new configuration option enables compiling pn532_uart as a module and
with this PN532 can be used with a simple UART to USB adapter.

Signed-off-by: Ziqian SUN (Zamir) <sztsian@gmail.com>