configs/fedora: Enable the interconnect SC7180 driver built-in
Enable the interconnect driver for the SC7180 SoC as built-in in order
to have output from the UART device console. As of v5.12 the QUP nodes
for Qualcomm boards has got their interconnect properties specified,
this means that the relevant interconnect provider needs to be builtin
for the UART device to probe and the console to be registered before
userspace needs to access it.

While here also make the pinctrl driver as built-in, as, apart from
being needed to have console output is needed to have early access to
different devices.

Note that a similar change might be also interested for SDM845
(following what the upstream config did)

Signed-off-by: Enric Balletbo i Serra <eballetbo@redhat.com>