[redhat] New configs in drivers/irqchip
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: QCOM_MPM [=n]
 Type  : tristate
 Defined at drivers/irqchip/Kconfig:433
   Prompt: QCOM MPM
   Depends on: ARCH_QCOM [=y]
   Location:
     Main menu
       -> Device Drivers
         -> IRQ chip support
 Selects: IRQ_DOMAIN_HIERARCHY [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>