[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FPROBE [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:250
   Prompt: Kernel Function Probe (fprobe)
   Depends on: FTRACE [=y] && FUNCTION_TRACER [=y] && DYNAMIC_FTRACE_WITH_REGS [=y] && HAVE_RETHOOK [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> Tracers (FTRACE [=y])
           -> Kernel Function Tracer (FUNCTION_TRACER [=y])
 Selects: RETHOOK [=y]

---

Cc: Jerome Marchand <jmarchan@redhat.com>
Cc: Joe Lawrence <joe.lawrence@redhat.com>
Cc: Josh Poimboeuf <jpoimboe@redhat.com>
Cc: Luis Claudio Goncalves <lgoncalv@redhat.com>
Cc: Waiman Long <longman@redhat.com>
Cc: Yannick Cote <ycote@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2: CONFIG_FPROBE enabled

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>