[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MARCH_Z16 [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:319
   Prompt: IBM z16
   Depends on: <choice> && y
   Location:
     Main menu
       -> Processor type and features
         -> Processor type (<choice> [=y])
 Selects: HAVE_MARCH_Z16_FEATURES [=n]

---

 Symbol: TUNE_Z16 [=n]
 Type  : bool
 Defined at arch/s390/Kconfig:390
   Prompt: IBM z16
   Depends on: <choice> && y
   Location:
     Main menu
       -> Processor type and features
         -> Tune code generation (<choice> [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>