[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_SM4_ARM64_CE_BLK [=n]
 Type  : tristate
 Defined at arch/arm64/crypto/Kconfig:56
   Prompt: SM4 in ECB/CBC/CFB/CTR modes using ARMv8 Crypto Extensions
   Depends on: CRYPTO [=y] && ARM64_CRYPTO [=y] && KERNEL_MODE_NEON [=y]
   Location:
     Main menu
       -> ARM64 Accelerated Cryptographic Algorithms (ARM64_CRYPTO [=y])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_SM4 [=n]

---

 Symbol: CRYPTO_SM4_ARM64_NEON_BLK [=n]
 Type  : tristate
 Defined at arch/arm64/crypto/Kconfig:62
   Prompt: SM4 in ECB/CBC/CFB/CTR modes using NEON instructions
   Depends on: CRYPTO [=y] && ARM64_CRYPTO [=y] && KERNEL_MODE_NEON [=y]
   Location:
     Main menu
       -> ARM64 Accelerated Cryptographic Algorithms (ARM64_CRYPTO [=y])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_SM4 [=n]

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>