redhat/Makefile: Fix dist-brew and distg-brew targets
The dist-brew and distg-brew targets are failing because the DIST variable
is defined as ".el9 DISTRO=rhel BUILD_TARGET="--scratch
rhel-9.1.0-test-pesign"".  This occurs because make doesn't allow the
setting of multiple variables on one line.

Break the variable assignments into separate lines, and fix the dist-brew
and distg-brew targets.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>