redhat/configs: Enable QAT devices for arches other than x86
Upstream 9c846c5d2d4e ("crypto: qat - Removes the x86 dependency on
the QAT drivers") removed the x86 dependency for the QAT drivers.
Enable QAT for the x86 only in CentOS Stream. Enable QAT for all the
supported architectures in Fedora. This means aarch64, ppc64le and
x86_64 as of now.

Signed-off-by: Vladis Dronov <vdronov@redhat.com>