redhat: properly handle binary files in patches
Pass --binary to git diff when generating the patch-*-redhat.patch in
genspec.sh, so that the patch is applicable also when there are diffs in
binary files.

Furthermore, since the `patch` utility doesn't support applying binary
diffs, switch the spec file to use `git apply` instead of `patch`. This
is OK, since the spec already has `BuildRequires: git-core`. In fact,
this means we can now remove `patch` from BuildRequires in favor of
`git-core`.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>