[redhat] New configs in sound/core
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_CTL_DEBUG [=n]
 Type  : bool
 Defined at sound/core/Kconfig:201
   Prompt: Enable debugging feature for control API
   Depends on: SOUND [=m] && !UML && SND [=m] && SND_DEBUG [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Sound card support (SOUND [=m])
           -> Advanced Linux Sound Architecture (SND [=m])

---

 Symbol: SND_CTL_INPUT_VALIDATION [=n]
 Type  : bool
 Defined at sound/core/Kconfig:191
   Prompt: Validate input data to control API
   Depends on: SOUND [=m] && !UML && SND [=m]
   Location:
     Main menu
       -> Device Drivers
         -> Sound card support (SOUND [=m])
           -> Advanced Linux Sound Architecture (SND [=m])

---

Cc: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_SND_CTL_DEBUG enabled for debug.
- CONFIG_SND_CTL_INPUT_VALIDATION enabled for debug.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>