Turn off some Kunit tests in pending
During the 6.0 merge window, 2 new commits have come in which cause major
issues with our kunit testing setup.

635dcd16844b08adcc1aa7a934893e47260619e4 for thunderbolt
fe5be808fa6c8df5cf939ec622514f970573bc88 for nitro_enclave
￼
In both cases, the modules have moved to the new implementation of
kunit_test_suite() in which the init is done in the main module itself
instead of as a separate kunit module.  This makes depmod fail because
the nitro_enclave and thunderbolt modules are both dependent upon
symbols that only exist in the mod-internal package.
There are multiple solutions to this, and Nico is looking at options, so
for now, let's just turn off these tests so things build while we
determine the best course.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>