[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_METEORLAKE [=n]
 Type  : tristate
 Defined at drivers/pinctrl/intel/Kconfig:154
   Prompt: Intel Meteor Lake pinctrl and GPIO driver
   Depends on: PINCTRL [=y] && (X86 [=y] || COMPILE_TEST [=n]) && ACPI [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Pin controllers (PINCTRL [=y])
           -> Intel pinctrl drivers
 Selects: PINCTRL_INTEL [=y]

---

Cc: Brian Masney <bmasney@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_PINCTRL_METEORLAKE moved under generic/x86.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>