common: enable INTEGRITY_MACHINE_KEYRING on all configuraitons
We need INTEGRITY_MACHINE_KEYRING for a number of edge use cases. It
allows use of user keys via the UEFI Machine Operator Key (MOK)
functionality and is useful for a number of things.

The requires SECONDARY_TRUSTED_KEYRING to be turned on in ARK.

Also includes a few cleanups on related configs.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>