enable efifb for Nvidia
The Nvidia prorietary driver has bugs that prevents it to be used with the
simpledrm driver. Namely, the fact that doesn't register a fbdev emulated
device and just relies on existing fbdev drivers to have a VT.

We have a workaround that skips a "simple-framebuffer" platform device if
the nvidia-drm.modeset=1 kernel cmdline param is set and "efi-framebuffer"
device is registered instead.

But for this to work, the legacy efifb and vesafb drivers must be enabled
and we only enabled them for x86. Which means that the issue would still
be present on Nvidia in other architectures such as aarch64.

Reported-by: Jon Nettleton <jon@solid-run.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>