[redhat] New configs in drivers/hwtracing
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HISI_PTT [=n]
 Type  : tristate
 Defined at drivers/hwtracing/ptt/Kconfig:2
   Prompt: HiSilicon PCIe Tune and Trace Device
   Depends on: (ARM64 [=y] || COMPILE_TEST [=n] && 64BIT [=y]) && PCI [=y] && HAS_DMA [=y] && HAS_IOMEM [=y] && PERF_EVENTS [=y]
   Location:
     Main menu
       -> Device Drivers
         -> HW tracing support

---

Cc: Doug Ledford <dledford@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>