[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_ARIA_AESNI_AVX_X86_64 [=n]
 Type  : tristate
 Defined at arch/x86/crypto/Kconfig:289
   Prompt: Ciphers: ARIA with modes: ECB, CTR (AES-NI/AVX/GFNI)
   Depends on: CRYPTO [=y] && !KMSAN [=n] && X86 [=y] && 64BIT [=y]
   Location:
     Main menu
       -> Cryptographic API (CRYPTO [=y])
         -> Accelerated Cryptographic Algorithms for CPU (x86)
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_SIMD [=y] && CRYPTO_ALGAPI [=y] && CRYPTO_ARIA [=n]

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>