[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IOMMUFD [=n]
 Type  : tristate
 Defined at drivers/iommu/iommufd/Kconfig:2
   Prompt: IOMMU Userspace API
   Depends on: IOMMU_SUPPORT [=y]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> IOMMU Userspace API (IOMMUFD [=n])
 Selects: INTERVAL_TREE [=y] && INTERVAL_TREE_SPAN_ITER [=n] && IOMMU_API [=y]

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Don Dutile <ddutile@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>