[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KVM_SMM [=y]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:121
   Prompt: System Management Mode emulation
   Depends on: VIRTUALIZATION [=y] && KVM [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> Kernel-based Virtual Machine (KVM) support (KVM [=m])
         -> System Management Mode emulation (KVM_SMM [=y])

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>