kernel.spec: add llvm-devel build requirement
The bpftool and the llvm system feature check requires the llvm-devel
package. The bpftool package is only created if selftests is defined and
when selftests is enabled, then samples/bpf will be built which will use
the llvm feature check. Therefore the llvm-devel package is added as a
build requirement of selftests to keep it consistent with the llvm
package already specified there.

A dependency of the llvm-devel package is llvm so we no longer need to
specify this package explicitly.

This change came in eb9d1acf634ba ('bpftool: Add LLVM as default
library for disassembling JIT-ed programs').

Signed-off-by: Scott Weaver <scweaver@redhat.com>