[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_TPMI [=n]
 Type  : tristate
 Defined at drivers/platform/x86/intel/Kconfig:185
   Prompt: Intel Topology Aware Register and PM Capsule Interface (TPMI)
   Depends on: X86_PLATFORM_DEVICES [=y] && INTEL_VSEC [=m] && X86_64 [=y]
   Location:
     -> Device Drivers
       -> X86 Platform Specific Device Drivers (X86_PLATFORM_DEVICES [=y])
         -> Intel Topology Aware Register and PM Capsule Interface (TPMI) (INTEL_TPMI [=n])

---

Cc: Mark Langsdorf <mlangsdo@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_INTEL_TPMI enabled as a module for x86

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>