redhat: fix duplicate jira issues in the resolves line
Bugzilla: INTERNAL
Upstream Status: RHEL only

The current genlog.py code does not check for duplicated jira issues
added to the resolves line. For bugzilla numbers, that check was present
when creating the resolved_bzs list.

To resolve it, convert everything to sets (which can't have duplicated
items), as such we don't need to do additional checks when printing the
resolves line. Also, make sure we sort not only the per changelog item
numbers, but also the items in the resolves line.

Fixes: 552fe563b55b ("redhat: add support for Jira issues in changelog")
Signed-off-by: Herton R. Krzesinski <herton@redhat.com>