[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_PANEL_AUO_A030JTN01 [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/panel/Kconfig:40
   Prompt: AUO A030JTN01
   Depends on: HAS_IOMEM [=y] && DRM [=y] && DRM_PANEL [=y] && SPI [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Panels
           -> AUO A030JTN01 (DRM_PANEL_AUO_A030JTN01 [=n])
 Selects: REGMAP_SPI [=m]

---

 Symbol: DRM_PANEL_ORISETECH_OTA5601A [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/panel/Kconfig:423
   Prompt: Orise Technology ota5601a RGB/SPI panel
   Depends on: HAS_IOMEM [=y] && DRM [=y] && DRM_PANEL [=y] && SPI [=y] && BACKLIGHT_CLASS_DEVICE [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Display Panels
           -> Orise Technology ota5601a RGB/SPI panel (DRM_PANEL_ORISETECH_OTA5601A [=n])
 Selects: REGMAP_SPI [=m]

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: "Michel Dänzer" <mdaenzer@redhat.com>
Cc: Karol Herbst <kherbst@redhat.com>
Cc: Ben Skeggs <bskeggs@redhat.com>
Cc: Jocelyn Falempe <jfalempe@redhat.com>
Cc: "Mika Penttilä" <mpenttil@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>