[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_AES_GCM_P10 [=m]
 Type  : tristate
 Defined at arch/powerpc/crypto/Kconfig:97
   Prompt: Stitched AES/GCM acceleration support on P10 or later CPU (PPC)
   Depends on: CRYPTO [=y] && !KMSAN [=n] && PPC [=y] && PPC64 [=y] && CPU_LITTLE_ENDIAN [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Accelerated Cryptographic Algorithms for CPU (powerpc)
         -> Stitched AES/GCM acceleration support on P10 or later CPU (PPC) (CRYPTO_AES_GCM_P10 [=m])
 Selects: CRYPTO_LIB_AES [=y] && CRYPTO_ALGAPI [=y] && CRYPTO_AEAD [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v1:
- CONFIG_CRYPTO_AES_GCM_P10 enabled as a built-in for rhel

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>