[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CSD_LOCK_WAIT_DEBUG_DEFAULT [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:1493
   Prompt: Default csd_lock_wait() debugging on at boot time
   Depends on: CSD_LOCK_WAIT_DEBUG [=y] && 64BIT [=y]
   Location:
     -> Kernel hacking
       -> Lock Debugging (spinlocks, mutexes, etc...)
         -> Debugging for csd_lock_wait(), called from smp_call_function*() (CSD_LOCK_WAIT_DEBUG [=y])
           -> Default csd_lock_wait() debugging on at boot time (CSD_LOCK_WAIT_DEBUG_DEFAULT [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>