Move imx8m configs from fedora to common
RHEl has gotten out of sync with kernel-ark IMX support so let's get
in sync. This is mostly a renaming of IMX-related fedora configs to
common configs. There are some other differences due to IMX9 not being
enabled in RHEL.

Verified that Fedora configs are identical before and after this patch.
Tested eln kernel on imx8mq-evk board.

Signed-off-by: Mark Salter <msalter@redhat.com>