CI: Remove legacy tag check cruft
Originally code was added to handle the script being repeatedly called
by detecting an existing tag and aborting to prevent a mess.

Over the years we removed the ark-patches, did more git merges and
strengthened dist-release.  This simplified the code paths and remove a
lot of fragile cases.

The original checks are unnecessary now that 'make dist-release' blocks
any changes from going in.  'make dist-release-tag' will also fail on
existing tags.

Remove the complicated check and rely on existing mechanisms.

In addition, ark-patches and its release structure died a couple of
years ago, remove that cruft too.  Instead just do the detached head for
ark-latest.

Also removed hardcoded 'os-build' and replaced with a variable in prep
for other branches to use this script (like rt and automotive).

Signed-off-by: Don Zickus <dzickus@redhat.com>