[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HID_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/hid/Kconfig:1304
   Prompt: KUnit tests for HID
   Depends on: HID_SUPPORT [=y] && HID [=y] && KUNIT [=m] && HID_BATTERY_STRENGTH [=y] && HID_UCLOGIC [=m]
   Visible if: HID_SUPPORT [=y] && HID [=y] && KUNIT [=m] && HID_BATTERY_STRENGTH [=y] && HID_UCLOGIC [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> HID bus support (HID_SUPPORT [=y])
         -> HID bus core support (HID [=y])
           -> Special HID drivers
             -> KUnit tests for HID (HID_KUNIT_TEST [=m])

---

 Symbol: HID_NVIDIA_SHIELD [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:791
   Prompt: NVIDIA SHIELD devices
   Depends on: HID_SUPPORT [=y] && HID [=y] && USB_HID [=y] && BT_HIDP [=m]
   Location:
     -> Device Drivers
       -> HID bus support (HID_SUPPORT [=y])
         -> HID bus core support (HID [=y])
           -> Special HID drivers
             -> NVIDIA SHIELD devices (HID_NVIDIA_SHIELD [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>