[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INPUT_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/input/Kconfig:169
   Prompt: KUnit tests for Input
   Depends on: INPUT [=y] && KUNIT [=m]
   Visible if: INPUT [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> KUnit tests for Input (INPUT_KUNIT_TEST [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>