[redhat] New configs in drivers/nvmem
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVMEM_IMX_OCOTP_ELE [=n]
 Type  : tristate
 Defined at drivers/nvmem/Kconfig:86
   Prompt: i.MX On-Chip OTP Controller support
   Depends on: NVMEM [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n]) && HAS_IOMEM [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> NVMEM Support (NVMEM [=y])
         -> i.MX On-Chip OTP Controller support (NVMEM_IMX_OCOTP_ELE [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>