Turn on NET_VENDOR_QUALCOMM for Fedora to enable rmnet
While rmnet will work with devices other than the ones gated behind
NET_VENDOR_QUALCOMM, the Kconfig for rmnet is also gated behind it. Turn
on NET_VENDOR_QUALCOMM for Fedora, but move the entries out of common
into RHEL so that configs are correct. Verified this makes no change to
the RHEL configs.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>