redhat: Fix UKI install with systemd >= 254
With systemd-254 'kernel-inspect' changed the output format and the
stop-gap copying script we have relies on the old one, i.e.

$ kernel-install inspect | grep KERNEL_INSTALL_ENTRY_TOKEN: | cut -d ' ' -f2`

Returns nothing. We could've changed this to:

$ kernel-install inspect | grep KERNEL_INSTALL_ENTRY_TOKEN | cut -d '=' -f2

to make things work again but luckily we don't need to: 'kernel-install' in
systemd-254 supports installing UKIs natively so our stop-gap can go for
good.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>