[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_CS42L43_I2C [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:245
   Prompt: Cirrus Logic CS42L43 (I2C)
   Depends on: HAS_IOMEM [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Cirrus Logic CS42L43 (I2C) (MFD_CS42L43_I2C [=n])
 Selects: REGMAP_I2C [=m] && MFD_CS42L43 [=n]

---

 Symbol: MFD_CS42L43_SDW [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:254
   Prompt: Cirrus Logic CS42L43 (SoundWire)
   Depends on: HAS_IOMEM [=y] && SOUNDWIRE [=m]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Cirrus Logic CS42L43 (SoundWire) (MFD_CS42L43_SDW [=n])
 Selects: REGMAP_SOUNDWIRE [=m] && MFD_CS42L43 [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>