[redhat] New configs in kernel/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DMA_NUMA_CMA [=y]
 Type  : bool
 Defined at kernel/dma/Kconfig:161
   Prompt: Enable separate DMA Contiguous Memory Area for NUMA Node
   Depends on: DMA_CMA [=y]
   Location:
     -> Library routines
       -> DMA Contiguous Memory Allocator (DMA_CMA [=y])
         -> Enable separate DMA Contiguous Memory Area for NUMA Node (DMA_NUMA_CMA [=y])

---

 Symbol: SWIOTLB_DYNAMIC [=n]
 Type  : bool
 Defined at kernel/dma/Kconfig:93
   Prompt: Dynamic allocation of DMA bounce buffers
   Depends on: SWIOTLB [=y]
   Location:
     -> Library routines
       -> Dynamic allocation of DMA bounce buffers (SWIOTLB_DYNAMIC [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>