[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_LOONGSON [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/loongson/Kconfig:3
   Prompt: DRM support for Loongson Graphics
   Depends on: HAS_IOMEM [=y] && DRM [=y] && PCI [=y] && MMU [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> DRM support for Loongson Graphics (DRM_LOONGSON [=n])
 Selects: DRM_KMS_HELPER [=y] && DRM_TTM [=m] && I2C [=y] && I2C_ALGOBIT [=m]

---

 Symbol: DRM_TTM_KUNIT_TEST [=n]
 Type  : tristate
 Defined at drivers/gpu/drm/Kconfig:198
   Prompt: KUnit tests for TTM
   Depends on: HAS_IOMEM [=y] && DRM [=y] && KUNIT [=m] && MMU [=y]
   Visible if: HAS_IOMEM [=y] && DRM [=y] && KUNIT [=m] && MMU [=y] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Graphics support
         -> KUnit tests for TTM (DRM_TTM_KUNIT_TEST [=n])
 Selects: DRM_TTM [=m] && DRM_EXPORT_FOR_TESTS [=y] && DRM_KUNIT_TEST_HELPERS [=m]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>