[redhat] New configs in fs/nls
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NLS_UCS2_UTILS [=m]
 Type  : tristate
 Defined at fs/nls/Kconfig:620
   Prompt: NLS UCS-2 UTILS
   Depends on: NLS [=y]
   Location:
     -> File systems
       -> Native language support (NLS [=y])
         -> NLS UCS-2 UTILS (NLS_UCS2_UTILS [=m])
 Selected by [m]:
   - CIFS [=m] && NETWORK_FILESYSTEMS [=y] && INET [=y]
 Selected by [n]:
   - JFS_FS [=n] && BLOCK [=y]
   - SMB_SERVER [=n] && NETWORK_FILESYSTEMS [=y] && INET [=y] && MULTIUSER [=y] && FILE_LOCKING [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>