[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IRSD200 [=n]
 Type  : tristate
 Defined at drivers/iio/proximity/Kconfig:35
   Prompt: Murata IRS-D200 PIR sensor
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Proximity and distance sensors
           -> Murata IRS-D200 PIR sensor (IRSD200 [=n])
 Selects: IIO_BUFFER [=y] && IIO_TRIGGERED_BUFFER [=m] && REGMAP_I2C [=m]

---

 Symbol: MCP4728 [=n]
 Type  : tristate
 Defined at drivers/iio/dac/Kconfig:392
   Prompt: MCP4728 DAC driver
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Digital to analog converters
           -> MCP4728 DAC driver (MCP4728 [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>