redhat: configs: generic: x86: Disable CONFIG_VIDEO_OV01A10 for x86 platform
On x86 the only devices with an OV01A10 sensor are devices with
the Intel IPU6 MIPI camera solution, which currently only works
with Intel's out of tree driver which is packaged for rpmfusion.

The new in kernel driver lacks integration with Intel's VSC
chip which sits between the sensor and the IPU6 and without this
things do not work.

So for now on these devices the ov01a10 module which comes
with the out of tree driver should be used instead.

Disabling the mainline version of the sensor driver should not
cause any regressions, since this sensor is only used together
with the IPU6 which needs an out of tree driver anyways.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>