[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SCMI_PERF_DOMAIN [=y]
 Type  : tristate
 Defined at drivers/firmware/arm_scmi/Kconfig:184
   Prompt: SCMI performance domain driver
   Depends on: ARM_SCMI_PROTOCOL [=y] || COMPILE_TEST [=n] && OF [=y]
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> ARM System Control and Management Interface Protocol
           -> SCMI performance domain driver (ARM_SCMI_PERF_DOMAIN [=y])
 Selects: PM_GENERIC_DOMAINS [=y]

---

 Symbol: QCOM_QSEECOM [=n]
 Type  : bool
 Defined at drivers/firmware/qcom/Kconfig:23
   Prompt: Qualcomm QSEECOM interface driver
   Depends on: QCOM_SCM [=y]=y
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> Qualcomm firmware drivers
           -> Qualcomm QSEECOM interface driver (QCOM_QSEECOM [=n])
 Selects: AUXILIARY_BUS [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARM_SCMI_PERF_DOMAIN move to arm/aarch64.
- CONFIG_QCOM_QSEECOM move to arm/aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>