[redhat] New configs in drivers/perf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DWC_PCIE_PMU [=n]
 Type  : tristate
 Defined at drivers/perf/Kconfig:220
   Prompt: Synopsys DesignWare PCIe PMU
   Depends on: PERF_EVENTS [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Performance monitor support
         -> Synopsys DesignWare PCIe PMU (DWC_PCIE_PMU [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>