[redhat] New configs in drivers/mtd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MTD_UBI_FAULT_INJECTION [=n]
 Type  : bool
 Defined at drivers/mtd/ubi/Kconfig:107
   Prompt: Fault injection capability of UBI device
   Depends on: MTD [=m] && MTD_UBI [=m] && FAULT_INJECTION_DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> Memory Technology Device (MTD) support (MTD [=m])
         -> Enable UBI - Unsorted block images (MTD_UBI [=m])
           -> Fault injection capability of UBI device (MTD_UBI_FAULT_INJECTION [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>