[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_VM_RB [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:822
   Prompt: Debug VM red-black trees
   Depends on: DEBUG_VM [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> Debug VM (DEBUG_VM [=y])
           -> Debug VM red-black trees (DEBUG_VM_RB [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>