[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_VF610 [=n]
 Type  : bool
 Defined at drivers/gpio/Kconfig:713
   Prompt: VF610 GPIO support
   Depends on: GPIOLIB [=y] && HAS_IOMEM [=y] && (ARCH_MXC [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Memory mapped GPIO drivers
           -> VF610 GPIO support (GPIO_VF610 [=n])
 Selects: GPIOLIB_IRQCHIP [=y]

 Commit: 7f2691a19627 (gpio: vf610: add gpiolib/IRQ chip driver for Vybrid)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>