[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_CONTPTE [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:2229
   Prompt: Contiguous PTE mappings for user memory
   Depends on: TRANSPARENT_HUGEPAGE [=y]
   Visible if: TRANSPARENT_HUGEPAGE [=y] && EXPERT [=y]
   Location:
     -> Kernel Features
       -> Contiguous PTE mappings for user memory (ARM64_CONTPTE [=y])

 Commit: 4602e5757bcc (arm64/mm: wire up PTE_CONT for user mappings)

---

 Symbol: ARM64_VA_BITS_52 [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig:1306
   Prompt: 52-bit
   Depends on: <choice> && (ARM64_PAN [=y] || !ARM64_SW_TTBR0_PAN [=n])
   Location:
     -> Kernel Features
       -> Virtual address space size (<choice> [=y])
         -> 52-bit (ARM64_VA_BITS_52 [=n])

 Commit: b6d00d47e81a (arm64: mm: Introduce 52-bit Kernel VAs)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>