[redhat] New configs in drivers/acpi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ACPI_APEI_EINJ_CXL [=y]
 Type  : bool
 Defined at drivers/acpi/apei/Kconfig:63
   Prompt: CXL Error INJection Support
   Depends on: ACPI [=y] && ACPI_APEI_EINJ [=m] && CXL_BUS [=m] && CXL_BUS [=m]<=ACPI_APEI_EINJ [=m]
   Location:
     -> Power management and ACPI options
       -> ACPI (Advanced Configuration and Power Interface) Support (ACPI [=y])
         -> APEI Error INJection (EINJ) (ACPI_APEI_EINJ [=m])
           -> CXL Error INJection Support (ACPI_APEI_EINJ_CXL [=y])

 Commit: 12fb28ea6b1c (EINJ: Add CXL error type support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>