[redhat] New configs in drivers/scsi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SCSI_PROTO_TEST [=m]
 Type  : tristate
 Defined at drivers/scsi/Kconfig:244
   Prompt: scsi_proto.h unit tests
   Depends on: SCSI [=y] && KUNIT [=m]
   Visible if: SCSI [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> SCSI device support
         -> scsi_proto.h unit tests (SCSI_PROTO_TEST [=m])

 Commit: 4977c0f4523e (scsi: scsi_proto: Add structures and constants related to I/O groups and streams)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>