Change fedora-stable-release.sh to use git am
When applying patches in fedora-stable release, we used to just use
patch.  While this worked well with things like config changes, it did
not work for patches to actual code because of the way we generate the
patch.  These need to be applied via git, or they are ignored.  Changing
this script to grab the current head commit on initial run, use git am
to apply patches per version, and git reset --hard to revert them after
a build fixe these issues.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>