[redhat] New configs in drivers/net/ethernet/microsoft
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MICROSOFT_MANA [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/microsoft/Kconfig:18
   Prompt: Microsoft Azure Network Adapter (MANA) support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MICROSOFT [=y] && PCI_MSI [=y] && (X86_64 || ARM64 [=y] && !CPU_BIG_ENDIAN [=n] && ARM64_4K_PAGES [=y]) && PCI_HYPERV [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Microsoft Network Devices (NET_VENDOR_MICROSOFT [=y])
             -> Microsoft Azure Network Adapter (MANA) support (MICROSOFT_MANA [=n])
 Selects: AUXILIARY_BUS [=y] && PAGE_POOL [=y]

 Commit: ca9c54d2d6a5 (net: mana: Add a driver for Microsoft Azure Network Adapter (MANA))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_MICROSOFT_MANA enable as a module for rhel-aarch64.
- CONFIG_MANA_INFINIBAND enable as a module for rhel-aarch64.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>