[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BASE_SMALL [=n]
 Type  : bool
 Defined at init/Kconfig:1588
   Prompt: Enable smaller-sized data structures for core
   Visible if: EXPERT [=y]
   Location:
     -> General setup
       -> Configure standard kernel features (expert users) (EXPERT [=y])
         -> Enable smaller-sized data structures for core (BASE_SMALL [=n])

 Commit: 27021649ec88 (printk: Remove redundant CONFIG_BASE_FULL)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>