[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FTRACE_VALIDATE_RCU_IS_WATCHING [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:977
   Prompt: Validate RCU is on during ftrace execution
   Depends on: FTRACE [=y] && FUNCTION_TRACER [=y] && ARCH_WANTS_NO_INSTR [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Validate RCU is on during ftrace execution (FTRACE_VALIDATE_RCU_IS_WATCHING [=n])

 Commit: b0e28a4b5bec (ftrace: make extra rcu_is_watching() validation check optional)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>