[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MEM_ALLOC_PROFILING [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:976
   Prompt: Enable memory allocation profiling
   Depends on: PROC_FS [=y] && !DEBUG_FORCE_WEAK_PER_CPU [=n]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> Enable memory allocation profiling (MEM_ALLOC_PROFILING [=n])
 Selects: CODE_TAGGING [=n] && PAGE_EXTENSION [=y] && SLAB_OBJ_EXT [=y]

 Commit: 22d407b164ff (lib: add allocation tagging support for memory allocation profiling)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>