[redhat] New configs in drivers/mailbox
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_MHU_V3 [=n]
 Type  : tristate
 Defined at drivers/mailbox/Kconfig:26
   Prompt: ARM MHUv3 Mailbox
   Depends on: MAILBOX [=y] && (HAS_IOMEM [=y] || COMPILE_TEST [=n]) && OF [=y]
   Location:
     -> Device Drivers
       -> Mailbox Hardware Support (MAILBOX [=y])
         -> ARM MHUv3 Mailbox (ARM_MHU_V3 [=n])

 Commit: ca1a8680b134 (mailbox: arm_mhuv3: Add driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>