usb: typec: qcom: add PMI8998 support
PMI8998 is the first generation of Qualcomm PMIC that features a type-c
controller. Many of the implementation details featured in the pm8150b
controller were originally included in PMI8998.

With the second generation (pm8150b), Qualcomm completely re-engineered
the register layout, moving all of the type-c registers to their own
block rather than interspersing them with the charger registers.

As a result, the two implementations are just barely different enough to
necessitate a new driver, since an additional layer of abstraction would
likely introduce more complexity than it saves.

PD support is left out for now since it is not supported on all
platforms and hasn't been tested yet.

Co-Developed-by: Jonas Dreßler <verdre@v0yd.nl>
Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>