gitlab-ci: remove unused RHMAINTAINERS variable
The one file in redhat/ that mentions this environment var RHMAINTAINERS
is gen_config_patches.sh. And it is only mentioned in a comment.

It looks like the actual code that used the var was removed long ago in
4f9393ca5c269989bf0a7e395e5cb939e43b974d ('Remove cc lines from
automatic configs').

With this, it should be safe to drop the setting.

Signed-off-by: Michael Hofmann <mhofmann@redhat.com>