[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PCI_PWRCTL_PWRSEQ [=n]
 Type  : tristate
 Defined at drivers/pci/pwrctl/Kconfig:8
   Prompt: PCI Power Control driver using the Power Sequencing subsystem
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI Power control drivers
           -> PCI Power Control driver using the Power Sequencing subsystem (PCI_PWRCTL_PWRSEQ [=n])
 Selects: POWER_SEQUENCING [=n] && PCI_PWRCTL [=n]

 Commit: 6140d185a43d (PCI/pwrctl: Add a PCI power control driver for power sequenced devices)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>