[redhat] New configs in drivers/net/ethernet/intel
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IDPF_SINGLEQ [=n]
 Type  : bool
 Defined at drivers/net/ethernet/intel/idpf/Kconfig:18
   Prompt: idpf singleq support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_INTEL [=y] && IDPF [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Intel devices (NET_VENDOR_INTEL [=y])
             -> Intel(R) Infrastructure Data Path Function Support (IDPF [=m])
               -> idpf singleq support (IDPF_SINGLEQ [=n])

 Commit: f771314d6b75 (idpf: compile singleq code only under default-n CONFIG_IDPF_SINGLEQ)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>