[redhat] New configs in drivers/net/ethernet/meta
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FBNIC [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/meta/Kconfig:20
   Prompt: Meta Platforms Host Network Interface
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_META [=y] && (X86_64 [=y] || COMPILE_TEST [=n]) && PCI_MSI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Meta Platforms devices (NET_VENDOR_META [=y])
             -> Meta Platforms Host Network Interface (FBNIC [=n])
 Selects: PHYLINK [=m]

 Commit: 546dd90be979 (eth: fbnic: Add scaffolding for Meta's NIC driver)

---

 Symbol: NET_VENDOR_META [=y]
 Type  : bool
 Defined at drivers/net/ethernet/meta/Kconfig:6
   Prompt: Meta Platforms devices
   Depends on: NETDEVICES [=y] && ETHERNET [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Meta Platforms devices (NET_VENDOR_META [=y])

 Commit: 546dd90be979 (eth: fbnic: Add scaffolding for Meta's NIC driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- NET_VENDOR_META disabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>