[redhat] New configs in drivers/platform
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CZNIC_PLATFORMS [=n]
 Type  : bool
 Defined at drivers/platform/cznic/Kconfig:7
   Prompt: Platform support for CZ.NIC's Turris hardware
   Location:
     -> Device Drivers
       -> Platform support for CZ.NIC's Turris hardware (CZNIC_PLATFORMS [=n])

 Commit: 992f1a3d4e88 (platform: cznic: Add preliminary support for Turris Omnia MCU)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>