[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: USERCOPY_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2805
   Prompt: KUnit Test for user/kernel boundary protections
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit Test for user/kernel boundary protections (USERCOPY_KUNIT_TEST [=m])

 Commit: cf6219ee889f (usercopy: Convert test_user_copy to KUnit test)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>