[redhat] New configs in net/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NET_FLOW_LIMIT [=y]
 Type  : bool
 Defined at net/Kconfig:359
   Prompt: Net flow limit
   Depends on: NET [=y] && RPS [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Net flow limit (NET_FLOW_LIMIT [=y])

 Commit: 99bbc7074190 (rps: selective flow shedding during softnet overflow)

---

 Symbol: RFS_ACCEL [=y]
 Type  : bool
 Defined at net/Kconfig:300
   Prompt: Hardware acceleration of RFS
   Depends on: NET [=y] && RPS [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Receive packet steering (RPS [=y])
           -> Hardware acceleration of RFS (RFS_ACCEL [=y])
 Selects: CPU_RMAP [=y]

 Commit: c445477d74ab (net: RPS: Enable hardware acceleration of RFS)

---

 Symbol: RPS [=y]
 Type  : bool
 Defined at net/Kconfig:292
   Prompt: Receive packet steering
   Depends on: NET [=y] && SMP [=y] && SYSFS [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Receive packet steering (RPS [=y])

 Commit: df3345457a7a (rps: add CONFIG_RPS)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>