gitlab-ci: restore bot pipeline behavior
For cki-ark-bot, the triggered MR pipelines are supposed to only build
the SRPM in the various build roots.

This got broken a while ago, most likely when kernel-ark was switched to
merge-result pipelines.

The previous implementation checked the git commit author, which doesn't
work for merged-result pipelines as the merge commit is generated by
GitLab. Switch that to check the name of the user who started the MR
pipeline, which should work in all cases.

Fixes https://gitlab.com/cki-project/kernel-ark/-/issues/163

Signed-off-by: Michael Hofmann <mhofmann@redhat.com>