[redhat] Forcibly disable Rust support for now
Counterintuitively, this is a prerequisite for fixing CKI so that Rust support
can be enabled, as the CI is external to the Linux Kernel source tree and has
to be updated out of band with a rust toolchain before we can turn on Rust support.

This is because if the toolchain is present before the configuration changes
to enable it is merged, then the builds fail because it wants CONFIG_RUST
configured when a rust toolchain is detected.

This is also needed for kernel branches that Rust support is not intended to
be enabled.

Signed-off-by: Neal Gompa <ngompa@fedoraproject.org>