fedora: disable RTL8192CU in Fedora
The rtl8xxxu driver supports all USB IDs that are supported in
the RTL8192CU so these drivers overlap and both load when a
device is used. This means that the one that happens to load
firsts is the one that "wins" and it's non deterministic so a
user may get different drivers on different boots and hence an
inconsistent user experience. Upstrea is planninng on removing
USB support in the rtlwifi driver and just use rtl8xxxu now
that the later is featue complete with AP support.

Signed-off-by: Peter Robinson <pbrobinson@redhat.com>