Add weakdep support to the kernel spec
Upstream change 61842868de13aa7fd7391c626e889f4d6f1450bf  when combined
with a modern kmod creates a situation where modules.weakdep is
"installed but not packaged".  We should handle this file like the rest
of the depmod generated files.  Do not package it, but create the ghost
entry for when it gets generated on post.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>