[redhat] New configs in drivers/net/can
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CAN_ROCKCHIP_CANFD [=n]
 Type  : tristate
 Defined at drivers/net/can/rockchip/Kconfig:3
   Prompt: Rockchip CAN-FD controller
   Depends on: NETDEVICES [=y] && CAN_DEV [=m] && CAN_NETLINK [=y] && (OF [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> CAN Device Drivers (CAN_DEV [=m])
           -> CAN device drivers with Netlink support (CAN_NETLINK [=y])
             -> Rockchip CAN-FD controller (CAN_ROCKCHIP_CANFD [=n])
 Selects: CAN_RX_OFFLOAD [=y]

 Commit: ff60bfbaf67f (can: rockchip_canfd: add driver for Rockchip CAN-FD controller)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>