[redhat] New configs in fs/erofs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: EROFS_FS_BACKED_BY_FILE [=y]
 Type  : bool
 Defined at fs/erofs/Kconfig:77
   Prompt: File-backed EROFS filesystem support
   Depends on: MISC_FILESYSTEMS [=y] && EROFS_FS [=m]
   Location:
     -> File systems
       -> Miscellaneous filesystems (MISC_FILESYSTEMS [=y])
         -> EROFS filesystem support (EROFS_FS [=m])
           -> File-backed EROFS filesystem support (EROFS_FS_BACKED_BY_FILE [=y])

 Commit: fb176750266a (erofs: add file-backed mount support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>