[redhat] New configs in security/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PROC_MEM_ALWAYS_FORCE [=y]
 Type  : bool
 Defined at security/Kconfig:35
   Prompt: Traditional /proc/pid/mem behavior
   Location:
     -> Security options
       -> Allow /proc/pid/mem access override (<choice> [=n])
         -> Traditional /proc/pid/mem behavior (PROC_MEM_ALWAYS_FORCE [=y])

 Commit: 41e8149c8892 (proc: add config & param to block forcing mem writes)

---

 Symbol: PROC_MEM_FORCE_PTRACE [=n]
 Type  : bool
 Defined at security/Kconfig:41
   Prompt: Require active ptrace() use for access override
   Location:
     -> Security options
       -> Allow /proc/pid/mem access override (<choice> [=n])
         -> Require active ptrace() use for access override (PROC_MEM_FORCE_PTRACE [=n])

 Commit: 41e8149c8892 (proc: add config & param to block forcing mem writes)

---

 Symbol: PROC_MEM_NO_FORCE [=n]
 Type  : bool
 Defined at security/Kconfig:47
   Prompt: Never
   Location:
     -> Security options
       -> Allow /proc/pid/mem access override (<choice> [=n])
         -> Never (PROC_MEM_NO_FORCE [=n])

 Commit: 41e8149c8892 (proc: add config & param to block forcing mem writes)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>