[redhat] New configs in fs/smb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CIFS_COMPRESSION [=n]
 Type  : bool
 Defined at fs/smb/client/Kconfig:207
   Prompt: SMB message compression (Experimental)
   Depends on: NETWORK_FILESYSTEMS [=y] && CIFS [=m]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> SMB3 and CIFS support (advanced network filesystem) (CIFS [=m])
           -> SMB message compression (Experimental) (CIFS_COMPRESSION [=n])

 Commit: d14bbfff259c (smb3: mark compression as CONFIG_EXPERIMENTAL and fix missing compression operation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>