[redhat] New configs in lib/xz
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: XZ_DEC_ARM64 [=y]
 Type  : bool
 Defined at lib/xz/Kconfig:33
   Prompt: ARM64 BCJ filter decoder
   Depends on: XZ_DEC [=y]
   Visible if: XZ_DEC [=y] && EXPERT [=y]
   Location:
     -> Library routines
       -> XZ decompression support (XZ_DEC [=y])
         -> ARM64 BCJ filter decoder (XZ_DEC_ARM64 [=y])
 Selects: XZ_DEC_BCJ [=y]

 Commit: 4b62813f5e7d (xz: Add ARM64 BCJ filter)

---

 Symbol: XZ_DEC_RISCV [=y]
 Type  : bool
 Defined at lib/xz/Kconfig:43
   Prompt: RISC-V BCJ filter decoder
   Depends on: XZ_DEC [=y]
   Visible if: XZ_DEC [=y] && EXPERT [=y]
   Location:
     -> Library routines
       -> XZ decompression support (XZ_DEC [=y])
         -> RISC-V BCJ filter decoder (XZ_DEC_RISCV [=y])
 Selects: XZ_DEC_BCJ [=y]

 Commit: 93d09773d1a5 (xz: add RISC-V BCJ filter)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>