[redhat] New configs in drivers/clk
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CLK_FIXED_RATE_KUNIT_TEST [=m]
 Type  : tristate
 Defined at drivers/clk/Kconfig:517
   Prompt: Basic fixed rate clk type KUnit test
   Depends on: COMMON_CLK [=y] && KUNIT [=m]
   Visible if: COMMON_CLK [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Device Drivers
       -> Common Clock Framework (COMMON_CLK [=y])
         -> Basic fixed rate clk type KUnit test (CLK_FIXED_RATE_KUNIT_TEST [=m])
 Selects: OF_OVERLAY [=n] && DTC [=y]

 Commit: 5776526beb95 (clk: Add KUnit tests for clk fixed rate basic type)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>