[redhat] New configs in drivers/ptp
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PTP_1588_CLOCK_VMCLOCK [=m]
 Type  : tristate
 Defined at drivers/ptp/Kconfig:134
   Prompt: Virtual machine PTP clock
   Depends on: (X86_TSC [=y] || ARM_ARCH_TIMER [=n]) && PTP_1588_CLOCK [=y] && ACPI [=y] && ARCH_SUPPORTS_INT128 [=y]
   Location:
     -> Device Drivers
       -> PTP clock support
         -> Virtual machine PTP clock (PTP_1588_CLOCK_VMCLOCK [=m])

 Commit: 205032724226 (ptp: Add support for the AMZNC10C 'vmclock' device)

---

 Symbol: PTP_S390 [=n]
 Type  : tristate
 Defined at drivers/ptp/Kconfig:244
   Prompt: S390 PTP driver
   Depends on: PTP_1588_CLOCK [=y] && S390 [=y]
   Location:
     -> Device Drivers
       -> PTP clock support
         -> S390 PTP driver (PTP_S390 [=n])

 Commit: 2d7de7a3010d (s390/time: Add PtP driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_PTP_S390 enable as a module for s390.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>