[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FAIL_SKB_REALLOC [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:2097
   Prompt: Fault-injection capability forcing skb to reallocate
   Depends on: FAULT_INJECTION_DEBUG_FS [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Fault-injection framework (FAULT_INJECTION [=y])
           -> Debugfs entries for fault-injection capabilities (FAULT_INJECTION_DEBUG_FS [=y])
             -> Fault-injection capability forcing skb to reallocate (FAIL_SKB_REALLOC [=n])

 Commit: 12079a59ce52 (net: Implement fault injection forcing skb reallocation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>