[redhat] New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PPC_FTRACE_OUT_OF_LINE_NUM_RESERVE [=32768]
 Type  : integer
 Defined at arch/powerpc/Kconfig:580
   Prompt: Number of ftrace out-of-line stubs to reserve within .text
   Depends on: PPC_FTRACE_OUT_OF_LINE [=y]
   Location:
     -> Kernel options
       -> Number of ftrace out-of-line stubs to reserve within .text (PPC_FTRACE_OUT_OF_LINE_NUM_RESERVE [=32768])

 Commit: cf9bc0efcce2 (powerpc64/ftrace: Support .text larger than 32MB with out-of-line stubs)

---

 Symbol: VPA_PMU [=n]
 Type  : tristate
 Defined at arch/powerpc/platforms/pseries/Kconfig:143
   Prompt: VPA PMU events
   Depends on: KVM_BOOK3S_64_HV [=m] && HV_PERF_CTRS [=y]
   Location:
     -> Platform support
       -> Hypervisor supplied PMU events (24x7 & GPCI) (HV_PERF_CTRS [=y])
         -> VPA PMU events (VPA_PMU [=n])

 Commit: 176cda0619b6 (powerpc/perf: Add perf interface to expose vpa counters)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>