[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LEDS_LP8864 [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:516
   Prompt: LED support for the TI LP8864/LP8866 4/6 channel LED drivers
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED support for the TI LP8864/LP8866 4/6 channel LED drivers (LEDS_LP8864 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: e14d879292b0 (leds: lp8864: Add support for Texas Instruments LP8864, LP8864S, LP8866 LED-backlights)

---

 Symbol: LEDS_ST1202 [=n]
 Type  : tristate
 Defined at drivers/leds/Kconfig:968
   Prompt: LED Support for STMicroelectronics LED1202 I2C chips
   Depends on: NEW_LEDS [=y] && LEDS_CLASS [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED Support for STMicroelectronics LED1202 I2C chips (LEDS_ST1202 [=n])
 Selects: LEDS_TRIGGERS [=y]

 Commit: 259230378c65 (leds: Add LED1202 I2C driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>