[redhat] New configs in net/unix
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AF_UNIX_OOB [=y]
 Type  : bool
 Defined at net/unix/Kconfig:19
   Prompt: UNIX: out-of-bound messages
   Depends on: NET [=y] && UNIX [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> Unix domain sockets (UNIX [=y])
           -> UNIX: out-of-bound messages (AF_UNIX_OOB [=y])

 Commit: 314001f0bf92 (af_unix: Add OOB support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>