[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_MSM8917 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/qcom/Kconfig.msm:140
   Prompt: Qualcomm 8917 pin controller driver
   Depends on: PINCTRL [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && PINCTRL_MSM [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Qualcomm core pin controller driver (PINCTRL_MSM [=y])
           -> Qualcomm 8917 pin controller driver (PINCTRL_MSM8917 [=n])

 Commit: ff5eb00255eb (pinctrl: qcom: Add MSM8917 tlmm pinctrl driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>