[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INT_SQRT_KUNIT_TEST [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:3188
   Prompt: Integer square root test
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> Integer square root test (INT_SQRT_KUNIT_TEST [=m])

 Commit: 0fafc9e1563f (lib/math: add int_sqrt test suite)

---

 Symbol: XARRAY_KUNIT [=m]
 Type  : tristate
 Defined at lib/Kconfig.debug:2459
   Prompt: KUnit test XArray code at runtime
   Depends on: RUNTIME_TESTING_MENU [=y] && KUNIT [=m]
   Visible if: RUNTIME_TESTING_MENU [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> KUnit test XArray code at runtime (XARRAY_KUNIT [=m])

 Commit: c7bb5cf9fc4e (xarray: port tests to kunit)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>