[redhat] New configs in drivers/regulator
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: REGULATOR_PF9453 [=n]
 Type  : tristate
 Defined at drivers/regulator/Kconfig:984
   Prompt: NXP PF9453 regulator driver
   Depends on: REGULATOR [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Voltage and Current Regulator Support (REGULATOR [=y])
         -> NXP PF9453 regulator driver (REGULATOR_PF9453 [=n])
 Selects: REGMAP_I2C [=m]

 Commit: 0959b6706325 (regulator: pf9453: add PMIC PF9453 support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>