[redhat] New configs in drivers/net/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: REALTEK_PHY_HWMON [=n]
 Type  : bool
 Defined at drivers/net/phy/realtek/Kconfig:9
   Prompt: HWMON support for Realtek PHYs
   Depends on: NETDEVICES [=y] && PHYLIB [=y] && REALTEK_PHY [=m] && HWMON [=y] && (REALTEK_PHY [=m]!=y [=y] || HWMON [=y]!=m [=m])
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> PHY Device support and infrastructure (PHYLIB [=y])
           -> Realtek PHYs (REALTEK_PHY [=m])
             -> HWMON support for Realtek PHYs (REALTEK_PHY_HWMON [=n])

 Commit: 33700ca45b7d (net: phy: realtek: add hwmon support for temp sensor on RTL822x)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>