[redhat] New configs in net/hsr
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PRP_DUP_DISCARD_KUNIT_TEST [=m]
 Type  : tristate
 Defined at net/hsr/Kconfig:44
   Prompt: PRP duplicate discard KUnit tests
   Depends on: NET [=y] && HSR [=m] && KUNIT [=m]
   Visible if: NET [=y] && HSR [=m] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> High-availability Seamless Redundancy (HSR & PRP) (HSR [=m])
           -> PRP duplicate discard KUnit tests (PRP_DUP_DISCARD_KUNIT_TEST [=m])

 Commit: 814dbf4b6c95 (net: hsr: Add KUnit test for PRP)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>