[redhat] New configs in drivers/cxl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CXL_FEATURES [=n]
 Type  : bool
 Defined at drivers/cxl/Kconfig:106
   Prompt: CXL: Features
   Depends on: CXL_BUS [=m] && CXL_PCI [=m]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> CXL (Compute Express Link) Devices Support (CXL_BUS [=m])
           -> CXL: Features (CXL_FEATURES [=n])

 Commit: f0e6a2329bf9 (cxl: Add Get Supported Features command for kernel usage)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>