[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NO_PAGE_MAPCOUNT [=n]
 Type  : bool
 Defined at mm/Kconfig:894
   Prompt: No per-page mapcount (EXPERIMENTAL)
   Depends on: TRANSPARENT_HUGEPAGE [=y]
   Location:
     -> Memory Management options
       -> Transparent Hugepage Support (TRANSPARENT_HUGEPAGE [=y])
         -> No per-page mapcount (EXPERIMENTAL) (NO_PAGE_MAPCOUNT [=n])

 Commit: e63ee43e3eda (mm: CONFIG_NO_PAGE_MAPCOUNT to prepare for not maintain per-page mapcounts in large folios)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>