[redhat] Move and consolidate arm configs
There are a fair number of arm specific configs in the rhel/generic
directory. They should really be in the rhel/generic/arm directory to
avoid polluting x86 with config options that don't apply.

But at the same time, a few of these configs are the same on fedora
and should be merged to the common/generic/arm directory.

Signed-off-by: Jeremy Linton <jlinton@redhat.com>