Flip GENKSYMS for RHEL
Upstream commit 9520a2b3f0b5e182f73410e45b9b92ea51d9b828 forces
GENDWARFKSYMS off if we have pahole 1.28 or 1.29 (even if the fix has
been backported to 1.29).  While we wait for 1.30 to release, we need to
flip GENKSYMS to avoid a mismatch.  This will be followed up with an
"include in release" MR that deletes this config again and reverts the
upstream commit which we can carry until pahole 1.30 is released.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>