[redhat] New configs in sound/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_HDA_ACPI [=n]
 Type  : tristate
 Defined at sound/pci/hda/Kconfig:45
   Prompt: HD Audio ACPI
   Depends on: SOUND [=m] && SND [=m] && ACPI [=y]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> HD-Audio
             -> HD Audio ACPI (SND_HDA_ACPI [=n])
 Selects: SND_HDA [=m]

 Commit: 4b214c9bbe26 (ALSA: hda - Add new driver for HDA controllers listed via ACPI)

Change for generic/x86 - SND_HDA_ACPI=m.

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>