[redhat] New configs in drivers/net/ethernet/huawei
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HINIC3 [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/huawei/hinic3/Kconfig:6
   Prompt: Huawei 3rd generation network adapters (HINIC3) support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_HUAWEI [=y] && !CPU_BIG_ENDIAN && (X86 [=y] || ARM64 || COMPILE_TEST [=n]) && PCI_MSI [=y] && 64BIT [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Huawei devices (NET_VENDOR_HUAWEI [=y])
             -> Huawei 3rd generation network adapters (HINIC3) support (HINIC3 [=n])
 Selects: AUXILIARY_BUS [=y] && PAGE_POOL [=y]

 Commit: 17fcb3dc12bb (hinic3: module initialization and tx/rx logic)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>