[redhat] New configs in net/sched
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NET_SCH_BPF [=n]
 Type  : bool
 Defined at net/sched/Kconfig:406
   Prompt: BPF-based Qdisc
   Depends on: NET [=y] && NET_SCHED [=y] && BPF_SYSCALL [=y] && BPF_JIT [=y] && DEBUG_INFO_BTF [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> QoS and/or fair queueing (NET_SCHED [=y])
           -> BPF-based Qdisc (NET_SCH_BPF [=n])

 Commit: c8240344956e (bpf: net_sched: Support implementation of Qdisc_ops in bpf)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>