[redhat] New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PHY_SNPS_EUSB2 [=n]
 Type  : tristate
 Defined at drivers/phy/Kconfig:46
   Prompt: SNPS eUSB2 PHY Driver
   Depends on: OF [=y] && (ARCH_EXYNOS [=n] || ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PHY Subsystem
         -> SNPS eUSB2 PHY Driver (PHY_SNPS_EUSB2 [=n])
 Selects: GENERIC_PHY [=y]

 Commit: 8d3b5f637546 (phy: move phy-qcom-snps-eusb2 out of its vendor sub-directory)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>