[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: USB_XHCI_SIDEBAND [=n]
 Type  : bool
 Defined at drivers/usb/host/Kconfig:107
   Prompt: xHCI support for sideband
   Depends on: USB_SUPPORT [=y] && USB [=y] && USB_XHCI_HCD [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> xHCI HCD (USB 3.0) support (USB_XHCI_HCD [=y])
           -> xHCI support for sideband (USB_XHCI_SIDEBAND [=n])

 Commit: de66754e9f80 (xhci: sideband: add initial api to register a secondary interrupter entity)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>