[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_SBRMI_I2C [=n]
 Type  : tristate
 Defined at drivers/misc/amd-sbi/Kconfig:2
   Prompt: AMD side band RMI support
   Depends on: I2C [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> AMD side band RMI support (AMD_SBRMI_I2C [=n])

 Commit: e15658676405 (hwmon/misc: amd-sbi: Move core sbrmi from hwmon to misc)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>