[redhat] New configs in sound/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SND_SOC_USB [=n]
 Type  : tristate
 Defined at sound/soc/Kconfig:94
   Prompt: SoC based USB audio offloading
   Depends on: SOUND [=m] && SND [=m] && SND_SOC [=m] && SND_USB_AUDIO [=m]
   Location:
     -> Device Drivers
       -> Sound card support (SOUND [=m])
         -> Advanced Linux Sound Architecture (SND [=m])
           -> ALSA for SoC audio support (SND_SOC [=m])
             -> SoC based USB audio offloading (SND_SOC_USB [=n])

 Commit: dba7759af789 (ASoC: Add SoC USB APIs for adding an USB backend)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>