kernel.spec: always provide kernel
Several packages include kernel as an installation requirement. In the
case of kernel-automotive and kernel-redhat (gemini) we need to
explicitly Provide the kernel package if these packages dependent on
kernel are included in a downstream distro. As an example, currently
automotive includes libguestfs, libguestfs-appliance and supermin in
their compose. Those require kernel.

When %{name} is kernel, the duplicate Provides does not currently cause
an RPM error and duplicate Provides are not reported by the final rpm
package. Therefore, there is no need to guard against including
duplicate Provides.

Signed-off-by: Scott Weaver <scweaver@redhat.com>