[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

    Symbol: CRYPTO_KRB5 [=n]
    Type  : tristate
    Defined at crypto/krb5/Kconfig:1
      Prompt: Kerberos 5 crypto
      Depends on: CRYPTO [=y]
      Location:
        -> Cryptographic API (CRYPTO [=y])
          -> Kerberos 5 crypto (CRYPTO_KRB5 [=n])
    Selects: CRYPTO_MANAGER [=y] && CRYPTO_KRB5ENC [=n] && CRYPTO_AUTHENC [=y] && CRYPTO_SKCIPHER [=y] && CRYPTO_HASH_INFO [=y] && CRYPTO_HMAC [=y] && CRYPTO_CMAC [=y] && CRYPTO_SHA1 [=y] && CRYPTO_SHA256 [=y] && CRYPTO_SHA512 [=y] && CRYPTO_CBC [=y] && CRYPTO_CTS [=y] && CRYPTO_AES [=y] && CRYPTO_CAMELLIA [=m]

    Commit: 3936f02bf2d3 (crypto/krb5: Implement Kerberos crypto core)

    Symbol: CRYPTO_KRB5ENC [=n]
    Type  : tristate
    Defined at crypto/Kconfig:237
      Prompt: Kerberos 5 combined hash+cipher support
      Depends on: CRYPTO [=y]
      Location:
        -> Cryptographic API (CRYPTO [=y])
          -> Crypto core or helper
            -> Kerberos 5 combined hash+cipher support (CRYPTO_KRB5ENC [=n])
Selects: CRYPTO_AEAD [=y] && CRYPTO_SKCIPHER [=y] && CRYPTO_MANAGER [=y] && CRYPTO_HASH [=y] && CRYPTO_NULL [=y]
Selected by [n]:
  - CRYPTO_KRB5 [=n] && CRYPTO [=y]

Commit: d1775a177f7f (crypto: Add 'krb5enc' hash and cipher AEAD algorithm)

Additionally, we need to set correct values for the s390 zfcpdump
configs as modules are not enabled there. And we need to change the
value of CONFIG_CRYPTO_SHA256_ARM64 as commit 6e36be511d284 now sets it
to 'default CRYPTO_LIB_SHA256'

---

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>