[redhat] New configs in drivers/irqchip
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ACLINT_SSWI [=n]
 Type  : bool
 Defined at drivers/irqchip/Kconfig:641
   Prompt: RISC-V ACLINT S-mode IPI Interrupt Controller
   Depends on: RISCV [=y] && SMP [=y]
   Location:
     -> Device Drivers
       -> IRQ chip support
         -> RISC-V ACLINT S-mode IPI Interrupt Controller (ACLINT_SSWI [=n])
 Selects: IRQ_DOMAIN_HIERARCHY [=y] && GENERIC_IRQ_IPI_MUX [=y]
 Selected by [n]:
   - THEAD_C900_ACLINT_SSWI [=n]

 ~~~
 Commit: df0f030ee7e4 (irqchip/thead-c900-aclint-sswi: Generalize aclint-sswi driver and add MIPS P800 support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>