[redhat] New configs in drivers/tty
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SERIAL_RSCI [=n]
 Type  : tristate
 Defined at drivers/tty/serial/Kconfig:678
   Prompt: Support for Renesas RZ/T2H SCI variant
   Depends on: TTY [=y] && HAS_IOMEM [=y] && SERIAL_SH_SCI [=y]
   Location:
     -> Device Drivers
       -> Character devices
         -> Enable TTY (TTY [=y])
           -> Serial drivers
             -> SuperH SCI(F) serial port support (SERIAL_SH_SCI [=y])
               -> Support for Renesas RZ/T2H SCI variant (SERIAL_RSCI [=n])

 ~~~
 Commit: 0666e3fe95ab (serial: sh-sci: Add support for RZ/T2H SCI)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>