[redhat] New configs in mm/damon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DAMON_STAT [=n]
 Type  : bool
 Defined at mm/damon/Kconfig:97
   Prompt: Build data access monitoring stat (DAMON_STAT)
   Depends on: DAMON_PADDR [=y]
   Location:
     -> Memory Management options
       -> Data Access Monitoring
         -> Build data access monitoring stat (DAMON_STAT) (DAMON_STAT [=n])

 ~~~
 Commit: 369c415e6073 (mm/damon: introduce DAMON_STAT module)

---

 ~~~
 Symbol: DAMON_STAT_ENABLED_DEFAULT [=n]
 Type  : bool
 Defined at mm/damon/Kconfig:105
   Prompt: Enable DAMON_STAT by default
   Depends on: DAMON_PADDR [=y]
   Location:
     -> Memory Management options
       -> Data Access Monitoring
         -> Enable DAMON_STAT by default (DAMON_STAT_ENABLED_DEFAULT [=n])

 ~~~
 Commit: 369c415e6073 (mm/damon: introduce DAMON_STAT module)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>