[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RMI4_F1A [=n]
 Type  : bool
 Defined at drivers/input/rmi4/Kconfig:85
   Prompt: RMI4 Function 1A (0D pointing)
   Depends on: INPUT [=y] && RMI4_CORE [=m]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Synaptics RMI4 bus support (RMI4_CORE [=m])
             -> RMI4 Function 1A (0D pointing) (RMI4_F1A [=n])

 ~~~
 Commit: 4619b6b97553 (Input: synaptics-rmi4 - add support for F1A)

---

 ~~~
 Symbol: RMI4_F21 [=n]
 Type  : bool
 Defined at drivers/input/rmi4/Kconfig:92
   Prompt: RMI4 Function 21 (PRESSURE)
   Depends on: INPUT [=y] && RMI4_CORE [=m]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Synaptics RMI4 bus support (RMI4_CORE [=m])
             -> RMI4 Function 21 (PRESSURE) (RMI4_F21 [=n])

 ~~~
 Commit: a9c95d17dc13 (Input: synaptics-rmi4 - add support for Forcepads (F21))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>