[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: OVMF_DEBUG_LOG [=n]
 Type  : bool
 Defined at drivers/firmware/efi/Kconfig:266
   Prompt: Expose OVMF firmware debug log via sysfs
   Depends on: EFI [=y]
   Location:
     -> Device Drivers
       -> Firmware Drivers
         -> EFI (Extensible Firmware Interface) Support
           -> Expose OVMF firmware debug log via sysfs (OVMF_DEBUG_LOG [=n])

 ~~~
 Commit: f393a761763c (efi: add ovmf debug log driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>