[redhat] New configs in arch/riscv
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ERRATA_MIPS [=n]
 Type  : bool
 Defined at arch/riscv/Kconfig.errata:24
   Prompt: MIPS errata
   Depends on: RISCV_ALTERNATIVE [=y]
   Location:
     -> CPU errata selection
       -> MIPS errata (ERRATA_MIPS [=n])

 ~~~
 Commit: 0b0ca959d206 (riscv: errata: Fix the PAUSE Opcode for MIPS P8700)

---

 ~~~
 Symbol: RISCV_ISA_VENDOR_EXT_MIPS [=y]
 Type  : bool
 Defined at arch/riscv/Kconfig.vendor:20
   Prompt: MIPS vendor extension support
   Location:
     -> Platform type
       -> Vendor extensions
         -> MIPS
           -> MIPS vendor extension support (RISCV_ISA_VENDOR_EXT_MIPS [=y])
 Selects: RISCV_ISA_VENDOR_EXT [=y]

 ~~~
 Commit: a8fed1bc03ac (riscv: Add xmipsexectl as a vendor extension)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>