[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARCH_LAN969X [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig.platforms:187
   Prompt: Microchip LAN969X SoC family
   Location:
     -> Platform selection
       -> Microchip SoC support
         -> Microchip LAN969X SoC family (ARCH_LAN969X [=n])
 Selects: PINCTRL [=y] && DW_APB_TIMER_OF [=n] && ARCH_MICROCHIP [=n]

 ~~~
 Commit: c8d4fbc7d2bf (arm64: lan969x: Add support for Microchip LAN969x SoC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>