[redhat] New configs in drivers/thermal
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RZG3E_THERMAL [=n]
 Type  : tristate
 Defined at drivers/thermal/renesas/Kconfig:38
   Prompt: Renesas RZ/G3E thermal driver
   Depends on: THERMAL [=y] && (ARCH_RENESAS [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> Thermal drivers (THERMAL [=y])
         -> Renesas RZ/G3E thermal driver (RZG3E_THERMAL [=n])

 ~~~
 Commit: 19d3a401a617 (thermal/drivers/renesas/rzg3e: Add thermal driver for the Renesas RZ/G3E SoC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>