[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PERSISTENT_HUGE_ZERO_FOLIO [=n]
 Type  : bool
 Defined at mm/Kconfig:798
   Prompt: Allocate a PMD sized folio for zeroing
   Depends on: TRANSPARENT_HUGEPAGE [=y]
   Location:
     -> Memory Management options
       -> Allocate a PMD sized folio for zeroing (PERSISTENT_HUGE_ZERO_FOLIO [=n])

 ~~~
 Commit: 2d8bd8049e89 (mm: add persistent huge zero folio)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>