[redhat] New configs in drivers/infiniband
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: INFINIBAND_IONIC [=n]
 Type  : tristate
 Defined at drivers/infiniband/hw/ionic/Kconfig:4
   Prompt: AMD Pensando DSC RDMA/RoCE Support
   Depends on: INFINIBAND [=m] && (INFINIBAND_USER_ACCESS [=m] || !INFINIBAND_USER_ACCESS [=m]) && !UML && NETDEVICES [=y] && ETHERNET [=y] && PCI [=y] && INET [=y] && IONIC [=m]
   Location:
     -> Device Drivers
       -> InfiniBand support (INFINIBAND [=m])
         -> AMD Pensando DSC RDMA/RoCE Support (INFINIBAND_IONIC [=n])

 ~~~
 Commit: 6603fbf158e9 (RDMA/ionic: Add Makefile/Kconfig to kernel build environment)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>