[redhat] New configs in drivers/crypto
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: CRYPTO_DEV_TI_DTHEV2 [=n]
 Type  : tristate
 Defined at drivers/crypto/ti/Kconfig:2
   Prompt: Support for TI DTHE V2 cryptography engine
   Depends on: CRYPTO [=y] && CRYPTO_HW [=y] && (ARCH_K3 [=y] || COMPILE_TEST [=n])
   Location:
     -> Cryptographic API (CRYPTO [=y])
       -> Hardware crypto devices (CRYPTO_HW [=y])
         -> Support for TI DTHE V2 cryptography engine (CRYPTO_DEV_TI_DTHEV2 [=n])
 Selects: CRYPTO_ENGINE [=m] && CRYPTO_SKCIPHER [=y] && CRYPTO_ECB [=y] && CRYPTO_CBC [=y]

 ~~~
 Commit: 52f641bc63a4 (crypto: ti - Add driver for DTHE V2 AES Engine (ECB, CBC))

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>