[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PRINTK_RINGBUFFER_KUNIT_TEST [=m]
 Type  : tristate
 Defined at init/Kconfig:1713
   Prompt: KUnit Test for the printk ringbuffer
   Depends on: PRINTK [=y] && KUNIT [=m]
   Visible if: PRINTK [=y] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     -> General setup
       -> Configure standard kernel features (expert users) (EXPERT [=y])
         -> Enable support for printk (PRINTK [=y])
           -> KUnit Test for the printk ringbuffer (PRINTK_RINGBUFFER_KUNIT_TEST [=m])

 ~~~
 Commit: 5ea2bcdfbf46 (printk: ringbuffer: Add KUnit test)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>