Enable CONFIG_VFIO_DEVICE_CDEV on Fedora
The CONFIG_VFIO_DEVICE_CDEV setting was initially left disabled
in Fedora with:

  commit 1b360fd559e31328f543e9c3e0c88accff4b5ff8
  Author: Justin M. Forbes <jforbes@fedoraproject.org>
  Date:   Fri Oct 13 16:07:43 2023 -0500

      Fedora configs for 6.6

A short while later it was set enabled in 'common/' with

  commit ec1c54cff1cde8be9439bc3b5ef067645e5158b0
  Author: Alex Williamson <alex@shazbot.org>
  Date:   Tue Nov 21 11:24:06 2023 -0700

      redhat/configs: Move IOMMUFD to common

      Enable resulting VFIO options

This overlooked the fact that Fedora had an explicit disable.

This removes the Fedora override so the common config takes
effect in Fedora.

This fixes the ability to use QEMU's recently added 'iommufd'
support in Fedora.

Reported-by: Cedric Le Goater <clg@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>