[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: DEBUG_BUGVERBOSE_DETAILED [=n]
 Type  : bool
 Defined at lib/Kconfig.debug:209
   Prompt: Verbose WARN_ON_ONCE() reporting (adds 100K)
   Visible if: DEBUG_BUGVERBOSE [=y]
   Location:
     -> Kernel hacking
       -> printk and dmesg options
         -> Verbose BUG() reporting (adds 70K) (DEBUG_BUGVERBOSE [=y])
           -> Verbose WARN_ON_ONCE() reporting (adds 100K) (DEBUG_BUGVERBOSE_DETAILED [=n])

 ~~~
 Commit: 687fac9d1b00 (bugs/core: Introduce the CONFIG_DEBUG_BUGVERBOSE_DETAILED Kconfig switch)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>