[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: ARM64_MPAM [=n]
 Type  : bool
 Defined at arch/arm64/Kconfig:2025
   Prompt: Enable support for MPAM
   Location:
     -> Kernel Features
       -> ARMv8.4 architectural features
         -> Enable support for MPAM (ARM64_MPAM [=n])
 Selects: ARM64_MPAM_DRIVER [=n] && ACPI_MPAM [=n]

 ~~~
 Commit: d8bf01d80919 (arm64: kconfig: Add Kconfig entry for MPAM)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_ARM64_MPAM: enable and move under arm/aarch64.
- CONFIG_ARM64_MPAM_DRIVER_DEBUG: disable.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>