[redhat] New configs in drivers/net/ethernet/mucse
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: MGBE [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/mucse/Kconfig:19
   Prompt: Mucse(R) 1GbE PCI Express adapters support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_MUCSE [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Mucse devices (NET_VENDOR_MUCSE [=y])
             -> Mucse(R) 1GbE PCI Express adapters support (MGBE [=n])

 ~~~
 Commit: ee61c10cd482 (net: rnpgbe: Add build support for rnpgbe)

---

 ~~~
 Symbol: NET_VENDOR_MUCSE [=y]
 Type  : bool
 Defined at drivers/net/ethernet/mucse/Kconfig:6
   Prompt: Mucse devices
   Depends on: NETDEVICES [=y] && ETHERNET [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Mucse devices (NET_VENDOR_MUCSE [=y])

 ~~~
 Commit: ee61c10cd482 (net: rnpgbe: Add build support for rnpgbe)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>