[redhat] New configs in security/selinux
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: SECURITY_SELINUX_AVC_HASH_BITS [=9]
 Type  : integer
 Range : [9 14]
 Defined at security/selinux/Kconfig:72
   Prompt: SELinux avc hashtable size
   Depends on: SECURITY_SELINUX [=y]
   Location:
     -> Security options
       -> SELinux Support (SECURITY_SELINUX [=y])
         -> SELinux avc hashtable size (SECURITY_SELINUX_AVC_HASH_BITS [=9])

 ~~~
 Commit: 641e02175861 (selinux: Introduce a new config to make avc cache slot size adjustable)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>