[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: GPIO_SHARED_PROXY [=m]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:2028
   Prompt: Proxy driver for non-exclusive GPIOs
   Depends on: GPIOLIB [=y] && (GPIO_SHARED [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Virtual GPIO drivers
           -> Proxy driver for non-exclusive GPIOs (GPIO_SHARED_PROXY [=m])
 Selects: AUXILIARY_BUS [=y]

 ~~~
 Commit: e992d54c6f97 (gpio: shared-proxy: implement the shared GPIO proxy driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>