[redhat] New configs in drivers/vfio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: XE_VFIO_PCI [=n]
 Type  : tristate
 Defined at drivers/vfio/pci/xe/Kconfig:2
   Prompt: VFIO support for Intel Graphics
   Depends on: VFIO [=m] && PCI [=y] && DRM_XE [=m] && PCI_IOV [=y]
   Location:
     -> Device Drivers
       -> VFIO Non-Privileged userspace driver framework (VFIO [=m])
         -> VFIO support for PCI devices
           -> VFIO support for Intel Graphics (XE_VFIO_PCI [=n])
 Selects: VFIO_PCI_CORE [=m]

 ~~~
 Commit: 1f5556ec8b9e (vfio/xe: Add device specific vfio_pci driver variant for Intel graphics)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>