[redhat] New configs in drivers/interconnect
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: INTERCONNECT_QCOM_KAANAPALI [=n]
 Type  : tristate
 Defined at drivers/interconnect/qcom/Kconfig:20
   Prompt: Qualcomm KAANAPALI interconnect driver
   Depends on: INTERCONNECT [=y] && INTERCONNECT_QCOM_RPMH_POSSIBLE [=y]
   Location:
     -> Device Drivers
       -> On-Chip Interconnect management support (INTERCONNECT [=y])
         -> Qualcomm KAANAPALI interconnect driver (INTERCONNECT_QCOM_KAANAPALI [=n])
 Selects: INTERCONNECT_QCOM_RPMH [=y] && INTERCONNECT_QCOM_BCM_VOTER [=y]

 ~~~
 Commit: c7f8ff611a14 (interconnect: qcom: add Kaanapali interconnect provider driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>