[redhat] New configs in drivers/nvdimm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RAMDAX [=m]
 Type  : tristate
 Defined at drivers/nvdimm/Kconfig:100
   Prompt: Support persistent memory interfaces on RAM carveouts
   Depends on: LIBNVDIMM [=m] && (X86_PMEM_LEGACY [=m] || OF [=n] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> NVDIMM (Non-Volatile Memory Device) Support (LIBNVDIMM [=m])
         -> Support persistent memory interfaces on RAM carveouts (RAMDAX [=m])

 ~~~
 Commit: 43bc0aa19a21 (nvdimm: allow exposing RAM carveouts as NVDIMM DIMM devices)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_RAMDAX: disable for rhel.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>